% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{anzsic_hs}
\alias{anzsic_hs}
\title{ABS AHECC Historic Conversion}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 9357 rows and 2 columns.
}
\usage{
anzsic_hs
}
\description{
ABS AHECC Historic Conversion
}
\keyword{datasets}
