% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{atlas_countries}
\alias{atlas_countries}
\title{Atlas of Economic Complexity Countries.}
\format{
Tibble with columns
\describe{
\item{location_code}{3-digit ISO country code}
\item{location_name_short_en}{English location name}
}
}
\usage{
atlas_countries
}
\description{
The name and ISO-3 representation of countries included in the Atlas of Economic Complexity
}
\keyword{datasets}
