% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{atlas_eci}
\alias{atlas_eci}
\title{Atlas of Economic Complexity Economic Complexity Index}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 6587 rows and 3 columns.
}
\usage{
atlas_eci
}
\description{
Atlas of Economic Complexity Economic Complexity Index
}
\keyword{datasets}
