% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{atlas_economic_complexity}
\alias{atlas_economic_complexity}
\title{Atlas of Economic Complexity Data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 4750834 rows and 6 columns.
}
\usage{
atlas_economic_complexity
}
\description{
Atlas of Economic Complexity Data
}
\keyword{datasets}
