% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{atlas_pci}
\alias{atlas_pci}
\title{Atlas of Economic Complexity Product Complexity Index}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 35897 rows and 3 columns.
}
\usage{
atlas_pci
}
\description{
Atlas of Economic Complexity Product Complexity Index
}
\keyword{datasets}
