% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{complexity_classification}
\alias{complexity_classification}
\title{Atlas of Economic Complexity Classification}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1217 rows and 3 columns.
}
\usage{
complexity_classification
}
\description{
Atlas of Economic Complexity Classification
}
\keyword{datasets}
