% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{complexity_colours}
\alias{complexity_colours}
\title{Atlas of Economic Complexity Colours}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 11 rows and 2 columns.
}
\usage{
complexity_colours
}
\description{
Atlas of Economic Complexity Colours
}
\keyword{datasets}
