% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sa2_indp1_occp1}
\alias{sa2_indp1_occp1}
\title{ABS Employment by Industry by Occupation (SA2, 1-digit Industry, 1-digit Occupation)}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 375896 rows and 4 columns.
}
\usage{
sa2_indp1_occp1
}
\description{
ABS Employment by Industry by Occupation (SA2, 1-digit Industry, 1-digit Occupation)
}
\keyword{datasets}
