% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sa3_indp1_occp1}
\alias{sa3_indp1_occp1}
\title{ABS Employment by Industry by Occupation (SA3, 1-digit Industry, 1-digit Occupation)}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 54568 rows and 4 columns.
}
\usage{
sa3_indp1_occp1
}
\description{
ABS Employment by Industry by Occupation (SA3, 1-digit Industry, 1-digit Occupation)
}
\keyword{datasets}
