% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sa3_indp4}
\alias{sa3_indp4}
\title{ABS Employment by Industry (SA3, 4-digit Industry)}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 257403 rows and 3 columns.
}
\usage{
sa3_indp4
}
\description{
ABS Employment by Industry (SA3, 4-digit Industry)
}
\keyword{datasets}
