% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sa3_occp4}
\alias{sa3_occp4}
\title{ABS Employment by Occupation (SA3, 4-digit Occupation)}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 161160 rows and 3 columns.
}
\usage{
sa3_occp4
}
\description{
ABS Employment by Occupation (SA3, 4-digit Occupation)
}
\keyword{datasets}
