% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{state_exports_data}
\alias{state_exports_data}
\title{Austrailan State and territory merchandise exports}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 191978 rows and 14 columns.
}
\usage{
state_exports_data
}
\description{
Austrailan State and territory merchandise exports
}
\details{
Placeholder
}
\keyword{datasets}
