% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{absemployment}
\alias{absemployment}
\alias{sa3_indp1_occp1}
\alias{sa2_indp1_occp1}
\alias{sa3_indp4}
\alias{sa3_occp4}
\alias{sa3_indp2_occp1}
\title{ABS Employment Data}
\format{
Tibble with columns

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 373008 rows and 3 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 244800 rows and 3 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 161160 rows and 3 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 285600 rows and 3 columns.
}
\usage{
sa3_indp1_occp1

sa2_indp1_occp1

sa3_indp4

sa3_occp4

sa3_indp2_occp1
}
\description{
ABS Employment Data

ABS Employment by Industry by Occupation (SA2, 1-digit Industry, 1-digit Occupation)

ABS Employment by Industry (SA3, 4-digit Industry)

ABS Employment by Occupation (SA3, 4-digit Occupation)
}
\keyword{datasets}
