% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{combined_exports}
\alias{combined_exports}
\title{Atlas of Economic Complexity and Australian State export data}
\format{
Tibble with columns:
\describe{
\item{year}{Year that product was exported.}
\item{location_code}{3 letter ISO country code.}
\item{hs_product_code}{4 digit HS1992 product code. }
\item{export_value}{Export value in US dollars.}
}
}
\source{
\url{https://dataverse.harvard.edu/dataverse/atlas}

\url{https://www.qgso.qld.gov.au/statistics/theme/economy/international-trade/exports}
}
\usage{
combined_exports
}
\description{
Atlas of Economic complexity data with Australia removed and replaced by
Australian State export data. State export data has been converted from the AHECC
system to the Harmonised System (1992 version)
}
\keyword{datasets}
