% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{state_economic_complexity}
\alias{state_economic_complexity}
\title{Australian State Economic Complexity data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 226289 rows and 8 columns.
}
\usage{
state_economic_complexity
}
\description{
Australian State Economic Complexity data
}
\keyword{datasets}
