% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{state_service_data}
\alias{state_service_data}
\title{Australian State services data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 920 rows and 4 columns.
}
\usage{
state_service_data
}
\description{
Australian State services data
}
\keyword{datasets}
